Abstracts classes panel
For re-use on the LocationPage
