+++ src/pages/classes.scss
@@ -25,23 +25,3 @@
     }
   }
 }
-
-.ClassPanel {
-  max-width: 1170px;
-  margin: auto auto 4rem;
-  width: 100%;
-  display: grid;
-  grid-template-columns: 1fr;
-
-  // Medium devices (tablets, 768px and up)
-  @media (min-width: 768px) {
-    grid-template-columns: 240px 1fr;
-    grid-gap: 1rem;
-  }
-
-  // Large devices (desktops, 992px and up)
-  @media (min-width: 992px) {
-    grid-template-columns: 320px 1fr;
-    grid-gap: 2rem;
-  }
-}
